% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.r
\name{create_and_draw_route_on_map}
\alias{create_and_draw_route_on_map}
\title{create_and_draw_route_on_map}
\usage{
create_and_draw_route_on_map(from, to)
}
\description{
create and draw route on map: an illustration of using decode_line_cpp()
}

