% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.r
\name{create_shortest_route_ft_visit}
\alias{create_shortest_route_ft_visit}
\title{create_shortest_route_ft_visit}
\usage{
create_shortest_route_ft_visit(from, to, visits = NULL, mode = c("driving",
  "walking", "bicycling", "transit"), method = "concorde", control = NULL,
  ...)
}
\arguments{
\item{from}{start address (text)}

\item{to}{ende address (text)}

\item{visits}{stops and visits address in between start and ende}
}
\value{
a named list of list(rftv, rftv_tbl)
 rftv: c(from, visits, to) address ordered by visit order in shortest route
 rftv_tbl: route data.frame in legs structure by following above rftv route
}
\description{
create shortest route from a start to an end visit several stops using TSP
}

