% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.r
\name{create_tsp_path}
\alias{create_tsp_path}
\alias{insert_dummy_node_tsp}
\title{create_tsp_path}
\usage{
create_tsp_path(m, mode = c("tsp", "atsp"), method = "concorde",
  control = NULL, ...)

insert_dummy_node_tsp(m)
}
\arguments{
\item{m}{dist matrix}

\item{mode}{tsp or atsp}

\item{method}{pass to TSP::solve_TSP()}

\item{control}{pass to TSP::solve_TSP()}

\item{...}{pass to TSP::solve_TSP()}
}
\value{
tsp path start from 1st node and ended to last node in dist matrix
}
\description{
create a tsp tour wrapper over TSP package

insert dummy node into a tsp/atsp dist matrix, subroutine of create_tsp_path.
}

