% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.r
\name{decode_line}
\alias{decode_line}
\title{decode_line}
\usage{
decode_line(encoded)
}
\description{
decode line from ggmap::route(from, to, structure = "route", output = "all")
wrapper over decode_line_cpp() and decode_line_r() depends on Rcpp available
}

