% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_line_cpp}
\alias{decode_line_cpp}
\title{decode_line_cpp}
\usage{
decode_line_cpp(encoded)
}
\description{
decode line return from ggmap::route(from, to, output = "all")
useful when draw route on map so route will follow real route.
}

