% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_line.r
\name{decode_line_r}
\alias{decode_line_r}
\title{decode_line_r}
\usage{
decode_line_r(encoded)
}
\description{
google polyline decoder borrowed from:
http://facstaff.unca.edu/mcmcclur/GoogleMaps/EncodePolyline/decode.js
}

