% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_ik.r
\name{ik_create_rgmp_icon_db}
\alias{ik_create_rgmp_icon_db}
\title{ik_create_rgmp_icon_db}
\usage{
ik_create_rgmp_icon_db()
}
\description{
create a rgmp_icon db for icon using in employee delivery application
icon: <icons code>-<color code>
 icons code (nm): offc - building, home - house, pkgs - package
 icons file (fl): offc - building, home - house, pkgs - package
 color code (cc): info, success, warning, danger, primary, purple and grey.
}

