% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_add_marker_on_maps}
\alias{mp_add_marker_on_maps}
\title{mp_add_marker_on_maps}
\usage{
mp_add_marker_on_maps(mm, dfmk)
}
\description{
map operation add marker in dfmk onto map rgmp
use mp_create_offc_marker, mp_create_home_marker, mp_create_pkgs_marker
}

