% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_add_route_on_maps}
\alias{mp_add_route_on_maps}
\title{mp_add_route_on_maps}
\usage{
mp_add_route_on_maps(mm, rltX, color = "#605ca8")
}
\arguments{
\item{rltX}{list of ordered address vector}
}
\value{
a map object with one continuous route for each list component
}
\description{
map operation add route at run-time onto map (server/client)
}

