% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_cgmp_home_id}
\alias{mp_create_cgmp_home_id}
\title{mp_create_cgmp_home_id}
\usage{
mp_create_cgmp_home_id(home_id, zoom = 12L)
}
\value{
cgmp:
 1. offc/home/ak_pkgs/rq_pkgs,
 2. current route to home include stops at accepted packages.
}
\description{
map operation: create client side map cgmp for home_id impersonated
}
\note{
we choose not to keep a global cgmp as we want to create carousel view in later version
}

