% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_cswb_cgmp_home_id}
\alias{mp_create_cswb_cgmp_home_id}
\title{mp_create_cswb_cgmp_home_id}
\usage{
mp_create_cswb_cgmp_home_id(input, output, session, home_id,
  outputId_cgmp = "cgmp")
}
\value{
a shinyUI widget
}
\description{
map operation: create client view (socialWidget(cgmp))
}

