% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_cswb_ctbl_home_id}
\alias{mp_create_cswb_ctbl_home_id}
\title{mp_create_cswb_ctbl_home_id}
\usage{
mp_create_cswb_ctbl_home_id(input, output, session, home_id,
  outputId_tb0DT = "ctb0", inputId_tb0_dl = "ctb0_dl",
  outputId_tb1DT = "ctb1", inputId_tb1_ak = "ctb1_ak",
  inputId_tb1_dk = "ctb1_dk")
}
\value{
a shinyUI widget
}
\description{
map operation: create client view (socialWidget(tb0DT + tb1DT))
}

