% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_cswb_home_id}
\alias{mp_create_cswb_home_id}
\title{mp_create_cswb_home_id}
\usage{
mp_create_cswb_home_id(home_id, ...)
}
\value{
a function wrapper over ygdashboard::socialWigetBox that take any shinyUI
component as ... input - simply a specific socialWigetBox for an employee
}
\description{
create social widget box for employee with home id impersonating
}
\note{
This is a nicer box that will be used to wrap over client single tbl and map
}

