% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_cswb_view_home_id}
\alias{mp_create_cswb_view_home_id}
\title{mp_create_cswb_view_home_id}
\usage{
mp_create_cswb_view_home_id(input, output, session, home_id,
  outputId_sw_ctbl = "sw_ctbl", outputId_sw_cgmp = "sw_cgmp",
  outputId_tb0DT = "ctb0", inputId_tb0_dl = "ctb0_dl",
  outputId_tb1DT = "ctb1", inputId_tb1_ak = "ctb1_ak",
  inputId_tb1_dk = "ctb1_dk", outputId_cgmp = "cgmp")
}
\value{
ctbl list (tb0, tb0DT, tb1, tb1DT) and cgmp object (cgmp)
}
\description{
map operation:
 1. create client view into output (outputId_sw_ctbl, outputId_sw_cgmp):
   1.1 ctbl view mp_create_cswb_ctbl_home_id (socialWidget(tb0DT + tb1DT)) into outputId_sw_ctbl
   1.2 cgmp view mp_create_cswb_cgmp_home_id (socialWidget(cgmp)) into outputId_sw_cgmp
 2. fulfill tb0DT, tb1DT, cgmp in client view created with
   1.1 ctbl (tb0DT, tb1DT) from mp_create_ctbl_home_id
   1.2 cgmp (cgmp) from mp_create_cgmp_home_id
}

