% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_cswb_view_home_id_static}
\alias{mp_create_cswb_view_home_id_static}
\title{mp_create_cswb_view_home_id_static}
\usage{
mp_create_cswb_view_home_id_static(home_id)
}
\value{
a static view
}
\description{
map operation:
 create static client view into output - plain html tagList
   1 ctbl view socialWidget(tb0DT + tb1DT) - fulfilled with tb0DT, tb1DT
   2 cgmp view socialWidget(cgmp) fufilled with cgmp
}
\note{
this is designed to be used to create carouselItem in carousel overview
}

