% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_ctbl_home_id}
\alias{mp_create_ctbl_home_id}
\title{mp_create_ctbl_home_id}
\usage{
mp_create_ctbl_home_id(home_id)
}
\value{
tb0: current route to home include stops at accepted packages
tb1: potential route to home include one additional stops at each received request
}
\description{
map operation: create client side tbl tb0 and tb1 for home_id impersonated
}

