% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_home_marker}
\alias{mp_create_home_marker}
\title{mp_create_home_marker}
\usage{
mp_create_home_marker(home_id)
}
\description{
create home marker df:
 mk: layerId,
 ik: icon file name, depends on route r-pkgs inst/icon
   and outside app folder www/img/<user-image.jpg/png>
 lat: latitude
 lng: longitude
 popup: popup info
}
\note{
choose to put user image in outside app folder www/img/ for easy extension
}

