% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_offc_marker}
\alias{mp_create_offc_marker}
\title{mp_create_offc_marker}
\usage{
mp_create_offc_marker(offc_id = NULL)
}
\description{
create offc marker df:
 mk: layerId,
 ik: icon file name (depends on route r-pkgs inst/icon)
 lat: latitude
 lng: longitude
 popup: popup info
}

