% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_create_rgmp_from_db}
\alias{mp_create_rgmp_from_db}
\title{mp_create_rgmp_from_db}
\usage{
mp_create_rgmp_from_db(offc_id = NULL, home_id = NULL, pkgs_id = NULL,
  showPopup = FALSE)
}
\description{
map operation: create server side map rgmp from offc/home/pkgs db
}

