% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_init_rgmp_obj}
\alias{mp_init_rgmp_obj}
\title{rgmp}
\usage{
mp_init_rgmp_obj()
}
\value{
rgmp object as a leaflet map object with icon and clickable popup
}
\description{
rgmp object hold global server leaflet map info
}

