% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_remove_rgmp_dl_pkgs}
\alias{mp_remove_rgmp_dl_pkgs}
\title{mp_remove_rgmp_dl_pkgs (animation)}
\usage{
mp_remove_rgmp_dl_pkgs(pkgs_id = NULL)
}
\description{
map operation: update server side map rgmp by removing package delivered
}

