% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_rgmp_home_ak_pkgs}
\alias{mp_rgmp_home_ak_pkgs}
\title{mp_rgmp_home_ak_pkgs (animation)}
\usage{
mp_rgmp_home_ak_pkgs(home_id, pkgs_id, zoom = 12L)
}
\value{
1. pkgs tbl and home tbl (side effect): update pkgs and home as request accepted
2. rgmp_ms (return_object): a transient rgmp with ployline and popup on request accepted
}
\description{
map operation: event response - home_id accept pkgs_id delivery request
 a wrap over rp_create_home_ak_pkgs()
}

