% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_rgmp_home_dk_pkgs}
\alias{mp_rgmp_home_dk_pkgs}
\title{mp_rgmp_home_dk_pkgs (animation)}
\usage{
mp_rgmp_home_dk_pkgs(home_id, pkgs_id, zoom = 12L)
}
\value{
1. pkgs tbl and home tbl (side effect): update pkgs and home as request declined
2. rgmp_ms (return_object): a transient rgmp with ployline and popup on request declined
}
\description{
map operation: event response - home_id decline pkgs_id delivery request
 a wrap over rp_create_home_dk_pkgs()
}

