% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_rgmp_home_dl_pkgs}
\alias{mp_rgmp_home_dl_pkgs}
\title{mp_rgmp_home_dl_pkgs (animation)}
\usage{
mp_rgmp_home_dl_pkgs(home_id, pkgs_id, zoom = 12L)
}
\value{
1. pkgs tbl and home tbl (side effect): update pkgs and home as request delivered
2. rgmp (side effect): update global rgmp with pkgs status change
3. rgmp_ms (return_object): a transient rgmp with ployline and popup on request delivered
}
\description{
map operation: event response - home_id deliver pkgs_id delivery request
 a wrap over rp_create_home_dk_pkgs()
}

