% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_rgmp_pkgs_rk_home_sdrqst}
\alias{mp_rgmp_pkgs_rk_home_sdrqst}
\title{mp_rgmp_pkgs_rk_home_sdrqst (animation)}
\usage{
mp_rgmp_pkgs_rk_home_sdrqst(pkgs_id, ns_rqst = 2L, zoom = 12L)
}
\value{
1. pkgs tbl and home tbl (side effect): update pkgs and home as request sent/sending
2. rgmp_ms (return_object): a transient rgmp with ployline and popup on request sent/sending
}
\description{
map operation: event response - pkgs_id send request to home_idlist
 a wrap over rp_create_pkgs_rk_home_sdrqst()
}

