% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_update_cgmp_home_pkgs}
\alias{mp_update_cgmp_home_pkgs}
\title{mp_update_cgmp_home_pkgs}
\usage{
mp_update_cgmp_home_pkgs(cgmp, home_id, pkgs_id)
}
\value{
cgmp:
 1. offc/home/ak_pkgs/rq_pkgs,
 2. current route to home include stops at accepted packages,
 3. potential route to home with one additonal stop at pkgs_id.
}
\description{
map operation: update client side map cgmp with potential route stopped at pkgs_id
}

