% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_update_rgmp_home_id}
\alias{mp_update_rgmp_home_id}
\title{mp_update_rgmp_home_id (animation)}
\usage{
mp_update_rgmp_home_id(home_id, zoom = 12L)
}
\value{
1. rgmp (side effect): updated global rgmp with new home_id
 2. rgmp_ms (return object): a transient rgmp with popup on new home_id
}
\description{
map operation: update server side map rgmp with new home_id
}

