% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_update_rgmp_offc_id}
\alias{mp_update_rgmp_offc_id}
\title{mp_update_rgmp_offc_id (animation)}
\usage{
mp_update_rgmp_offc_id(offc_id, zoom = 12L)
}
\value{
1. rgmp (side effect): updated global rgmp with new offc_id
 2. rgmp_ms (return object): a transient rgmp with popup on new offc_id
}
\description{
map operation: update server side map rgmp with new offc_id
}

