% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_update_rgmp_pkgs_id}
\alias{mp_update_rgmp_pkgs_id}
\title{mp_update_rgmp_pkgs_id (animation)}
\usage{
mp_update_rgmp_pkgs_id(pkgs_id, zoom = 12L)
}
\value{
1. rgmp (side effect): updated global rgmp with new pkgs_id
2. rgmp_ms (return object): a transient rgmp with popup on new pkgs_id
}
\description{
map operation: update server side map rgmp with new pkgs_id
}

