% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_mp.r
\name{mp_update_rgmp_rlth_id}
\alias{mp_update_rgmp_rlth_id}
\title{mp_update_rgmp_rlth_id (animation)}
\usage{
mp_update_rgmp_rlth_id(home_id, pkgs_pq_id = 0L, zoom = 12L,
  color = "#605ca8", mm = NULL)
}
\value{
1. rgmp (side effect): no side effect on rgmp
2. rgmp_ms (return object): a transient rgmp with current route to home
}
\description{
map operation: update server side map rgmp with (current/potential) rlth_id
}
\note{
we won't use mp_update_rgmp_rlth_id to update global rgmp like we did in offc/home/pkgs
 so we take an additional input mm as leaflet map object as we expect to add route info
 on local/transient map where some marker/popup is shown
}

