% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_home_ak_pkgs}
\alias{rp_create_home_ak_pkgs}
\title{rp_create_home_ak_pkgs}
\usage{
rp_create_home_ak_pkgs(home_id, pkgs_id)
}
\arguments{
\item{home_id:}{int scalar}

\item{pkgs_id:}{int vector}
}
\value{
status code: 0L ok!, 1L home_id not receive deliver request for pkgs_id
}
\description{
operation on server side as response to event from client side:
 an employee (home_id) accept (ak) a package (pkgs_id) delivery request first
}
\note{
empoloyee is the first employee who accepts delivery request of this package
}

