% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_home_dk_pkgs}
\alias{rp_create_home_dk_pkgs}
\title{rp_create_home_dk_pkgs}
\usage{
rp_create_home_dk_pkgs(home_id, pkgs_id)
}
\arguments{
\item{home_id:}{int scalar}

\item{pkgs_id:}{int vector}
}
\value{
status code:
 0L ok!
 1L home_id not receive deliver request for pkgs_id
 2L pkgs_id all request sent earlier has been declined
}
\description{
operation on server side as response to event from client side:
 an employee (home_id) decline (dk) a package (pkgs_id) delivery request
 decline before anyone accept - might trigger server sending request [?]
 if so, server should check has all request sent been declined and decide
}

