% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_home_dl_pkgs}
\alias{rp_create_home_dl_pkgs}
\title{rp_create_home_dl_pkgs}
\usage{
rp_create_home_dl_pkgs(home_id, pkgs_id)
}
\arguments{
\item{home_id:}{int scalar}

\item{pkgs_id:}{int vector}
}
\value{
status code:
 0L ok!
 1L home_id not accept deliver request for some pkgs_id, or
    pkgs_id not all accepted being confirmed as delivered.
}
\description{
operation on server side as response to event from client side:
 an employee (home_id) delivered (dl) a package (pkgs_id) accepted earlier
allow deliver all accepted packages at once in current version,
as a simulation of confirmation of delivery when arrived back home,
should allow confirm on delivery one package at a time in later version
}
\note{
param home_id is enough in current version, as pkgs_id should always be aid,
still keep pkgs_id as a param for a later extension on confirm pkgs_id being
delivered one at a time
}

