% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_offc_id}
\alias{rp_create_offc_id}
\title{rp_create_offc_id}
\usage{
rp_create_offc_id(id, nm, vd)
}
\value{
status code:
 0L ok!
 1L id exists,
 2L NA latlon.
}
\description{
create an offc id record into offc tbl:
 1. in situ update global offc tbl
}

