% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_offc_rt_home}
\alias{rp_create_offc_rt_home}
\title{rp_create_offc_rt_home}
\usage{
rp_create_offc_rt_home(home_id)
}
\arguments{
\item{home_id:}{home id employee}
}
\value{
home_id rlth tbl (current + potential)
}
\description{
create route from offc to home with stops at package accepted
 (+ one requested) for delivery, calculate shortest route by TSP
}
\note{
one of the most often used function in application
}

