% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_offc_rt_home_krnt}
\alias{rp_create_offc_rt_home_krnt}
\title{rp_create_offc_rt_home_krnt}
\usage{
rp_create_offc_rt_home_krnt(home_id)
}
\arguments{
\item{home_id:}{home id employee}
}
\value{
status code: 0L ok!
}
\description{
create route from offc to home with stops at package accepted
 for delivery in an earlier time, calculate shortest route by TSP
}
\note{
a wrap over rp_create_offc_rt_home_single
a special case of rp_create_offc_rt_home, with current route only,
 update on rlth tbl directly and return status code instead of rlth tbl
}

