% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_offc_rt_home_single}
\alias{rp_create_offc_rt_home_single}
\title{rp_create_offc_rt_home_single}
\usage{
rp_create_offc_rt_home_single(offc_id, home_id, pkgs_id, pkgs_pq_id = 0L)
}
\value{
rlth tb
}
\description{
subroutine of rp_create_offc_rt_home
create a signle route from offc to home with stops at packages pkgs_id
 and optional pkgs_pq_id
if pkgs_pq_id = 0L:
 create current route without additional delivery request
if pkgs_pq_id > 0L:
 create potential route with one additional at pkgs_pq_id
}

