% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_pkgs_id}
\alias{rp_create_pkgs_id}
\title{rp_create_pkgs_id}
\usage{
rp_create_pkgs_id(id, vd, fid)
}
\value{
status code:
 0L ok!
 1L id exists,
 2L NA latlon,
 3L fid not exists.
}
\description{
create an pkgs id record into pkgs tbl:
 1. in situ update global pkgs tbl
}

