% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_create_pkgs_rk_home_sdrqst}
\alias{rp_create_pkgs_rk_home_sdrqst}
\title{rp_create_pkgs_rk_home_sdrqst}
\usage{
rp_create_pkgs_rk_home_sdrqst(pkgs_id, ns_rqst = 2L)
}
\value{
status code: 0L ok!, 1L request has been sent to all home_id in this offc_id.
}
\description{
create pkgs a ranked home_idlist and send deliver request to top ns_rqst
 unsent home_id, using rp_create_pkgs_rk_home_idlist()
}

