% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_init_home_tbl}
\alias{rp_init_home_tbl}
\title{home}
\usage{
rp_init_home_tbl(loadInitTbl = TRUE)
}
\value{
home tbl in data.table
 id: int scalar: home id
 fn: chr string: employee first name
 ln: chr string: employee last name
 vd: chr string: home address
 md: chr string: commute mode
   driving, walking, bicycling, transit
 img: chr string: path to image file
 lat: num scalar: latitude
 lng: num scalar: longitude
 fid: int scalar: employee's office id
   office id must exist when insert an employee in such office
 pid: int single list: pending request(s) of package id(s)
 aid: int single list: accepted request(s) of package id(s)
 did: int single list: declined request(s) of package id(s)
 lid: int single list: delivered request(s) of package id(s)
}
\description{
home tbl hold employee info
}

