% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_init_offc_tbl}
\alias{rp_init_offc_tbl}
\title{offc}
\usage{
rp_init_offc_tbl(loadInitTbl = TRUE)
}
\value{
offc tbl in data.table
 id: int scalar: office id
 nm: chr string: office name
 vd: chr string: office address
 lat: num scalar: latitude
 lng: num scalar: longitude
}
\description{
offc tbl hold office info
}

