% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_rp.r
\name{rp_init_pkgs_tbl}
\alias{rp_init_pkgs_tbl}
\title{pkgs}
\usage{
rp_init_pkgs_tbl(loadInitTbl = TRUE)
}
\value{
pkgs tbl in data.table
 id: int scalar: package id
 vd: chr string: package address
 lat: num scalar: latitude
 lng: num scalar: longitude
 fid: int scalar: package's office id
   office id must exist when insert a package in such office
 kid: int single list: ranked home id for sending request [x]
   kid is removed from pkgs tbl as rank should be determined
   at runtime depends on other packages accepted by employee
 hid: int single list: home id delivery request has been sent
 aid: int scalar: employee home id accepted delivery request first
   aid will be cleaned after delivery, move into lid
 did: int single list: employee home id declined delivery request
 lid: int scalar: employee home id delivered package should be aid
   lid should be package id previouly delivered and will include aid
   lid should be cleaned after a period of time for performance issue
   and in that case we should create another column for tracking n_lid
 sid: int scalar: package status code:
   0L open, 1L accepted, and 2L delivered
}
\description{
pkgs tbl hold package info
}

